Fixed reverse loop crash

+++ src/sacn/source_state.c
@@ -208,8 +208,8 @@ int process_sources(process_sources_behavior_t behavior)
 {
   int num_sources_tracked = 0;
 
-  // Iterate the sources backwards to allow for removals
-  for (size_t i = get_num_sources() - 1; i >= 0; --i)
+  // Iterate the sources backwards to allow for removals (i must not be unsigned)
+  for (int i = get_num_sources() - 1; i >= 0; --i)
   {
     SacnSource* source = get_source(i);
 
@@ -251,8 +251,8 @@ void process_universe_discovery(SacnSource* source)
 // Needs lock
 void process_universes(SacnSource* source)
 {
-  // Iterate the universes backwards to allow for removals
-  for (size_t i = source->num_universes - 1; i >= 0; --i)
+  // Iterate the universes backwards to allow for removals (i must not be unsigned)
+  for (int i = source->num_universes - 1; i >= 0; --i)
   {
     SacnSourceUniverse* universe = &source->universes[i];
 
@@ -270,8 +270,8 @@ void process_universes(SacnSource* source)
 // Needs lock
 void process_unicast_dests(SacnSource* source, SacnSourceUniverse* universe)
 {
-  // Iterate unicast destinations backwards to allow for removals
-  for (size_t i = universe->num_unicast_dests - 1; i >= 0; --i)
+  // Iterate unicast destinations backwards to allow for removals (i must not be unsigned)
+  for (int i = universe->num_unicast_dests - 1; i >= 0; --i)
   {
     SacnUnicastDestination* dest = &universe->unicast_dests[i];
